requirements.txt
spotipy==2.24.0
google-generativeai==0.8.4
python-dotenv==1.0.1
